PBA Progress
For the PBA, changed the canvas and applied 2 map functions: one for the mouth that moves with the vocals and one for the eye size that moves with the drum. I also did a teeth function that moves in the same way as the mouth, that is following the voice.